# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "server project",
        "title": "Swagger server API",
        "contact": {
            "name": "antgo",
            "url": "https://github.com/small-ek/antgo",
            "email": "56494565@qq.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9001",
    "basePath": "/api",
    "paths": {
...
        "/sys-admin-users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员用户"
                ],
                "summary": "获取管理员用户分页数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current_page",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "boolean"
                        },
                        "collectionFormat": "csv",
                        "description": "排序方式 true 降序 false 升序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "电子邮件",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "过滤条件",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "标识",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nick_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "忽略字段",
                        "name": "omit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "排序字段",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页显示数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "选择字段",
                        "name": "select",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态:1=成功;2=失败",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Write"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/models.SysAdminUsers"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员用户"
                ],
                "summary": "创建管理员用户数据",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysAdminUsersRequestForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            }
        },
        "/sys-admin-users/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员用户"
                ],
                "summary": "获取管理员用户详情数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Write"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SysAdminUsers"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员用户"
                ],
                "summary": "修改管理员用户数据",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysAdminUsersRequestForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员用户"
                ],
                "summary": "删除管理员用户数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/response.Write"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.SysAdminUsers": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "电子邮件",
                    "type": "string"
                },
                "id": {
                    "description": "标识",
                    "type": "integer"
                },
                "nick_name": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机",
                    "type": "string"
                },
                "status": {
                    "description": "状态:1=成功;2=失败",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "request.SysAdminUsersRequestForm": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "电子邮件",
                    "type": "string"
                },
                "id": {
                    "description": "标识",
                    "type": "integer"
                },
                "nick_name": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机",
                    "type": "string"
                },
                "status": {
                    "description": "状态:1=成功;2=失败",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "response.Page": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "Data set of the current page 数据集合"
                },
                "total": {
                    "description": "Total pages 总数量",
                    "type": "integer"
                }
            }
        },
        "response.Write": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Status Code 业务逻辑的状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 实际数据"
                },
                "error": {
                    "description": "Error message 错误相关的信息"
                },
                "message": {
                    "description": "Msg Prompt message 简短描述",
                    "type": "string"
                }
            }
        },
...
    }
}